Some reorganisation of chez generation
For an output file 'f', now generate a subdirectory in the executable
directory 'f_app', and put 'f.so' and 'f.ss' in there, along with any C
libraries we might have needed. Then generate a shell script 'f' to set
the library path to `f_app` and invoke `f.so`.

In :exec mode, f = _tmpchez, so we're using the executable directory for
the temporary files.

This means all the libraries and code needed to run the generated
program are copied to a single relocatable target directory.

This almost certainly won't work on Windows. Please help! Thanks :)