Small improvements in codegen performance
Lots of time is taken in finding the definitions to compile, mostly
because the definition itself needs to be decoded from the binary format
in the ttc, but we can at least make the rest of the process fast.

Inlining takes a while - we could improve this later by inlining as we
go, per file, rather than doing it just before compilation.