[ refactor ] remove redundant check isLet
The `isLet` check is essentially the disjunction of the `Maybe Bool`
and the function looking up the value to check whether it is a `Let`.
But:

* `getBinder` does the same lookup & then dispatches on Let
* the non-let cases in both branches of the if-then-else are equal

So we can actually get rid of that `isLet` and it should compute the
same result. We still use the `Maybe Bool` to potentially avoid running
`getBinder` if we already know the local value is not a let.